+++ packages/fast-components-react-base/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.25.4](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-components-react-base@4.25.3...@microsoft/fast-components-react-base@4.25.4) (2020-03-25)
+
+**Note:** Version bump only for package @microsoft/fast-components-react-base
+
+
+
+
+
 ## [4.25.3](https://github.com/Microsoft/fast-dna/compare/@microsoft/fast-components-react-base@4.25.2...@microsoft/fast-components-react-base@4.25.3) (2020-03-25)
 
 **Note:** Version bump only for package @microsoft/fast-components-react-base